+++ packages/stream_chat/test/src/db/chat_persistence_client_test.dart
@@ -2,6 +2,8 @@ import 'package:stream_chat/src/core/api/requests.dart';
 import 'package:stream_chat/src/core/api/sort_order.dart';
 import 'package:stream_chat/src/core/models/channel_model.dart';
 import 'package:stream_chat/src/core/models/channel_state.dart';
+import 'package:stream_chat/src/core/models/draft.dart';
+import 'package:stream_chat/src/core/models/draft_message.dart';
 import 'package:stream_chat/src/core/models/event.dart';
 import 'package:stream_chat/src/core/models/filter.dart';
 import 'package:stream_chat/src/core/models/member.dart';
@@ -56,6 +58,10 @@ class TestPersistenceClient extends ChatPersistenceClient {
   Future<void> deletePollVotesByPollIds(List<String> pollIds) => Future.value();
 
   @override
+  Future<void> deleteDraftMessageByCid(String cid, {String? parentId}) =>
+      Future.value();
+
+  @override
   Future<void> disconnect({bool flush = false}) => throw UnimplementedError();
 
   @override
@@ -97,6 +103,18 @@ class TestPersistenceClient extends ChatPersistenceClient {
 
   @override
   Future<List<Read>> getReadsByCid(String cid) async => [];
+
+  @override
+  Future<Draft?> getDraftMessageByCid(
+    String cid, {
+    String? parentId,
+  }) async =>
+      Draft(
+        channelCid: cid,
+        parentId: parentId,
+        createdAt: DateTime.now(),
+        message: DraftMessage(id: 'message-id', text: 'message-text'),
+      );
 
   @override
   Future<List<Message>> getReplies(String parentId,
@@ -152,6 +170,9 @@ class TestPersistenceClient extends ChatPersistenceClient {
 
   @override
   Future<void> updatePolls(List<Poll> polls) => Future.value();
+
+  @override
+  Future<void> updateDraftMessages(List<Draft> draftMessages) => Future.value();
 }
 
 void main() {
@@ -192,6 +213,21 @@ void main() {
     test('updatePolls', () async {
       final poll = Poll(id: 'poll-id', name: 'poll-name', options: const []);
       persistenceClient.updatePolls([poll]);
+    });
+
+    test('deleteDraftMessageByCid', () {
+      const cid = 'test:cid';
+      const parentId = 'parent-id';
+      persistenceClient.deleteDraftMessageByCid(cid, parentId: parentId);
+    });
+
+    test('updateDraftMessages', () async {
+      final draft = Draft(
+        channelCid: 'test:cid',
+        createdAt: DateTime.now(),
+        message: DraftMessage(id: 'message-id', text: 'message-text'),
+      );
+      persistenceClient.updateDraftMessages([draft]);
     });
 
     test('updateChannelThreads', () async {