+++ packages/stream_chat_flutter/lib/src/bottom_sheets/edit_message_sheet.dart
@@ -143,6 +143,7 @@ class _EditMessageSheetState extends State<EditMessageSheet> {
                       ),
                     );
                   },
+                  elevation: 0,
                   messageInputController: controller,
                   // Disallow editing poll for now as it's not supported.
                   allowedAttachmentPickerTypes: [