+++ packages/stream_chat_flutter/lib/src/localization/translations.dart
@@ -446,8 +446,14 @@ abstract class Translations {
   /// The label for "Enter your comment".
   String get enterYourCommentLabel;
 
+  /// The confirmation text shown when the user tries to end a poll.
+  String get endVoteConfirmationText;
+
   /// The label for "Create".
   String get createLabel;
+
+  /// The label for "End".
+  String get endLabel;
 
   /// The label for Poll voting mode.
   ///
@@ -538,6 +544,9 @@ abstract class Translations {
 
   /// The text for poll when someone created
   String pollSomeoneCreatedText(String username);
+
+  /// The label for draft message
+  String get draftLabel;
 }
 
 /// Default implementation of Translation strings for the stream chat widgets
@@ -1100,7 +1109,14 @@ Attachment limit exceeded: it's not possible to add more than $limit attachments
   String get enterYourCommentLabel => 'Enter your comment';
 
   @override
+  String get endVoteConfirmationText =>
+      'Are you sure you want to end the vote?';
+
+  @override
   String get createLabel => 'Create';
+
+  @override
+  String get endLabel => 'End';
 
   @override
   String pollVotingModeLabel(PollVotingMode votingMode) {
@@ -1204,4 +1220,7 @@ Attachment limit exceeded: it's not possible to add more than $limit attachments
 
   @override
   String pollSomeoneCreatedText(String username) => '$username created';
+
+  @override
+  String get draftLabel => 'Draft';
 }