+++ packages/stream_chat_flutter/lib/src/message_input/stream_message_input.dart
@@ -6,12 +6,11 @@ import 'package:flutter/services.dart';
 import 'package:stream_chat_flutter/platform_widget_builder/src/platform_widget_builder.dart';
 import 'package:stream_chat_flutter/src/message_input/attachment_button.dart';
 import 'package:stream_chat_flutter/src/message_input/command_button.dart';
-import 'package:stream_chat_flutter/src/message_input/dm_checkbox.dart';
-import 'package:stream_chat_flutter/src/message_input/quoted_message_widget.dart';
+import 'package:stream_chat_flutter/src/message_input/dm_checkbox_list_tile.dart';
 import 'package:stream_chat_flutter/src/message_input/quoting_message_top_area.dart';
-import 'package:stream_chat_flutter/src/message_input/simple_safe_area.dart';
 import 'package:stream_chat_flutter/src/message_input/stream_message_input_icon_button.dart';
 import 'package:stream_chat_flutter/src/message_input/tld.dart';
+import 'package:stream_chat_flutter/src/misc/simple_safe_area.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
 const _kCommandTrigger = '/';
@@ -517,6 +516,8 @@ class StreamMessageInputState extends State<StreamMessageInput>
       ..addListener(_onChangedDebounced);
   }
 
+  StreamSubscription<Draft?>? _draftStreamSubscription;
+
   @override
   void initState() {
     super.initState();
@@ -528,18 +529,42 @@ class StreamMessageInputState extends State<StreamMessageInput>
     _effectiveFocusNode.addListener(_focusNodeListener);
 
     WidgetsBinding.instance.endOfFrame.then((_) {
-      if (!mounted) return;
-
-      // Call the listener once to make sure the initial state is reflected
-      // correctly in the UI.
-      _onChangedDebounced.call();
-
-      // Resumes the cooldown if the channel has currently an active cooldown.
-      if (!_isEditing) {
-        final channel = StreamChannel.of(context).channel;
-        _effectiveController.startCooldown(channel.getRemainingCooldown());
-      }
+      if (mounted) return _initializeState();
     });
+  }
+
+  void _initializeState() {
+    // Call the listener once to make sure the initial state is reflected
+    // correctly in the UI.
+    _onChangedDebounced.call();
+
+    final channel = StreamChannel.of(context).channel;
+    final config = StreamChatConfiguration.of(context);
+
+    // Resumes the cooldown if the channel has currently an active cooldown.
+    if (!_isEditing) {
+      _effectiveController.startCooldown(channel.getRemainingCooldown());
+    }
+
+    // Starts listening to the draft stream for the current channel/thread.
+    if (!_isEditing && config.draftMessagesEnabled) {
+      final draftStream = switch (_effectiveController.message.parentId) {
+        final parentId? => channel.state?.threadDraftStream(parentId),
+        _ => channel.state?.draftStream,
+      };
+
+      _draftStreamSubscription = draftStream?.distinct().listen(_onDraftUpdate);
+    }
+  }
+
+  void _onDraftUpdate(Draft? draft) {
+    // If the draft is removed, reset the controller.
+    if (draft == null) return _effectiveController.reset();
+
+    // Otherwise, update the controller with the draft message.
+    if (draft.message case final draftMessage) {
+      _effectiveController.message = draftMessage.toMessage();
+    }
   }
 
   @override
@@ -585,171 +610,173 @@ class StreamMessageInputState extends State<StreamMessageInput>
 
   @override
   Widget build(BuildContext context) {
+    bool canSendOrUpdateMessage(List<ChannelCapability> capabilities) {
+      var result = capabilities.contains(ChannelCapability.sendMessage);
+      if (_isEditing) {
+        result |= capabilities.contains(ChannelCapability.updateOwnMessage);
+        result |= capabilities.contains(ChannelCapability.updateAnyMessage);
+      }
+
+      return result;
+    }
+
     final channel = StreamChannel.of(context).channel;
-    if (channel.state != null && !channel.canSendMessage) {
-      return SafeArea(
-        child: Padding(
-          padding: const EdgeInsets.symmetric(
-            horizontal: 24,
-            vertical: 15,
+    final messageInput = BetterStreamBuilder(
+      stream: channel.ownCapabilitiesStream.map(canSendOrUpdateMessage),
+      initialData: canSendOrUpdateMessage(channel.ownCapabilities),
+      builder: (context, enabled) {
+        // Allow the user to send messages if the user has the permission to
+        // send messages or if the user is editing a message.
+        if (enabled) {
+          return _buildAutocompleteMessageInput(context);
+        }
+
+        // Otherwise, show the no permission message.
+        return _buildNoPermissionMessage(context);
+      },
+    );
+
+    final shadow = widget.shadow ?? _messageInputTheme.shadow;
+    final elevation = widget.elevation ?? _messageInputTheme.elevation;
+    return Material(
+      elevation: elevation ?? 8,
+      child: DecoratedBox(
+        decoration: BoxDecoration(
+          color: _messageInputTheme.inputBackgroundColor,
+          boxShadow: [if (shadow != null) shadow],
+        ),
+        child: SimpleSafeArea(
+          enabled: widget.enableSafeArea ?? _messageInputTheme.enableSafeArea,
+          child: Center(child: messageInput),
+        ),
+      ),
+    );
+  }
+
+  Widget _buildAutocompleteMessageInput(BuildContext context) {
+    return StreamAutocomplete(
+      focusNode: _effectiveFocusNode,
+      messageEditingController: _effectiveController,
+      fieldViewBuilder: _buildMessageInput,
+      autocompleteTriggers: [
+        ...widget.customAutocompleteTriggers,
+        StreamAutocompleteTrigger(
+          trigger: _kCommandTrigger,
+          triggerOnlyAtStart: true,
+          optionsViewBuilder: (
+            context,
+            autocompleteQuery,
+            messageEditingController,
+          ) {
+            final query = autocompleteQuery.query;
+            return StreamCommandAutocompleteOptions(
+              query: query,
+              channel: StreamChannel.of(context).channel,
+              onCommandSelected: (command) {
+                _effectiveController.command = command.name;
+                // removing the overlay after the command is selected
+                StreamAutocomplete.of(context).closeSuggestions();
+              },
+            );
+          },
+        ),
+        if (widget.enableMentionsOverlay)
+          StreamAutocompleteTrigger(
+            trigger: _kMentionTrigger,
+            optionsViewBuilder: (
+              context,
+              autocompleteQuery,
+              messageEditingController,
+            ) {
+              final query = autocompleteQuery.query;
+              return StreamMentionAutocompleteOptions(
+                query: query,
+                channel: StreamChannel.of(context).channel,
+                mentionAllAppUsers: widget.mentionAllAppUsers,
+                mentionsTileBuilder: widget.userMentionsTileBuilder,
+                onMentionUserTap: (user) {
+                  // adding the mentioned user to the controller.
+                  _effectiveController.addMentionedUser(user);
+
+                  // accepting the autocomplete option.
+                  StreamAutocomplete.of(context)
+                      .acceptAutocompleteOption(user.name);
+                },
+              );
+            },
           ),
-          child: Text(
-            context.translations.sendMessagePermissionError,
-            style: _messageInputTheme.inputTextStyle,
-          ),
+      ],
+    );
+  }
+
+  Widget _buildMessageInput(
+    BuildContext context,
+    StreamMessageEditingController controller,
+    FocusNode focusNode,
+  ) {
+    return StreamMessageValueListenableBuilder(
+      valueListenable: controller,
+      builder: (context, value, _) => Padding(
+        padding: const EdgeInsets.all(8),
+        child: Column(
+          spacing: 8,
+          mainAxisSize: MainAxisSize.min,
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: <Widget?>[
+            _buildTopMessageArea(context),
+            _buildTextField(context),
+            _buildDmCheckbox(context),
+          ].nonNulls.toList(),
+        ),
+      ),
+    );
+  }
+
+  Widget? _buildTopMessageArea(BuildContext context) {
+    if (_hasQuotedMessage && !_isEditing) {
+      // Ensure this doesn't show on web & desktop
+      return PlatformWidgetBuilder(
+        mobile: (context, child) => child,
+        child: QuotingMessageTopArea(
+          hasQuotedMessage: _hasQuotedMessage,
+          onQuotedMessageCleared: widget.onQuotedMessageCleared,
         ),
       );
     }
 
-    return StreamMessageValueListenableBuilder(
-      valueListenable: _effectiveController,
-      builder: (context, value, _) {
-        Widget child = DecoratedBox(
-          decoration: BoxDecoration(
-            color: _messageInputTheme.inputBackgroundColor,
-            boxShadow: widget.shadow == null
-                ? (_streamChatTheme.messageInputTheme.shadow == null
-                    ? []
-                    : [_streamChatTheme.messageInputTheme.shadow!])
-                : [widget.shadow!],
-          ),
-          child: SimpleSafeArea(
-            enabled: widget.enableSafeArea ??
-                _streamChatTheme.messageInputTheme.enableSafeArea ??
-                true,
-            child: GestureDetector(
-              onPanUpdate: (details) {
-                if (details.delta.dy > 0) {
-                  _effectiveFocusNode.unfocus();
-                }
-              },
-              child: Column(
-                mainAxisSize: MainAxisSize.min,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  if (_hasQuotedMessage && !_isEditing)
-                    // Ensure this doesn't show on web & desktop
-                    PlatformWidgetBuilder(
-                      mobile: (context, child) => child,
-                      child: QuotingMessageTopArea(
-                        hasQuotedMessage: _hasQuotedMessage,
-                        onQuotedMessageCleared: widget.onQuotedMessageCleared,
-                      ),
-                    )
-                  else if (_effectiveController.ogAttachment != null)
-                    OGAttachmentPreview(
-                      attachment: _effectiveController.ogAttachment!,
-                      onDismissPreviewPressed: () {
-                        _effectiveController.clearOGAttachment();
-                        _effectiveFocusNode.unfocus();
-                      },
-                    ),
-                  Padding(
-                    padding: const EdgeInsets.symmetric(vertical: 8),
-                    child: _buildTextField(context),
-                  ),
-                  if (_effectiveController.message.parentId != null &&
-                      !widget.hideSendAsDm)
-                    Padding(
-                      padding: const EdgeInsets.only(
-                        right: 12,
-                        left: 12,
-                        bottom: 10,
-                      ),
-                      child: DmCheckbox(
-                        foregroundDecoration: BoxDecoration(
-                          border: _effectiveController.showInChannel
-                              ? null
-                              : Border.all(
-                                  color: _streamChatTheme
-                                      .colorTheme.textHighEmphasis
-                                      // ignore: deprecated_member_use
-                                      .withOpacity(0.5),
-                                  width: 2,
-                                ),
-                          borderRadius: BorderRadius.circular(3),
-                        ),
-                        color: _effectiveController.showInChannel
-                            ? _streamChatTheme.colorTheme.accentPrimary
-                            : _streamChatTheme.colorTheme.barsBg,
-                        onTap: () {
-                          _effectiveController.showInChannel =
-                              !_effectiveController.showInChannel;
-                        },
-                        crossFadeState: _effectiveController.showInChannel
-                            ? CrossFadeState.showFirst
-                            : CrossFadeState.showSecond,
-                      ),
-                    ),
-                ],
-              ),
-            ),
-          ),
-        );
-        if (!_isEditing) {
-          child = Material(
-            elevation: widget.elevation ??
-                _streamChatTheme.messageInputTheme.elevation ??
-                8,
-            color: _messageInputTheme.inputBackgroundColor,
-            child: child,
-          );
-        }
+    if (_effectiveController.ogAttachment != null) {
+      return OGAttachmentPreview(
+        attachment: _effectiveController.ogAttachment!,
+        onDismissPreviewPressed: () {
+          _effectiveController.clearOGAttachment();
+          _effectiveFocusNode.unfocus();
+        },
+      );
+    }
 
-        return StreamAutocomplete(
-          focusNode: _effectiveFocusNode,
-          messageEditingController: _effectiveController,
-          fieldViewBuilder: (_, __, ___) => child,
-          autocompleteTriggers: [
-            ...widget.customAutocompleteTriggers,
-            StreamAutocompleteTrigger(
-              trigger: _kCommandTrigger,
-              triggerOnlyAtStart: true,
-              optionsViewBuilder: (
-                context,
-                autocompleteQuery,
-                messageEditingController,
-              ) {
-                final query = autocompleteQuery.query;
-                return StreamCommandAutocompleteOptions(
-                  query: query,
-                  channel: StreamChannel.of(context).channel,
-                  onCommandSelected: (command) {
-                    _effectiveController.command = command.name;
-                    // removing the overlay after the command is selected
-                    StreamAutocomplete.of(context).closeSuggestions();
-                  },
-                );
-              },
-            ),
-            if (widget.enableMentionsOverlay)
-              StreamAutocompleteTrigger(
-                trigger: _kMentionTrigger,
-                optionsViewBuilder: (
-                  context,
-                  autocompleteQuery,
-                  messageEditingController,
-                ) {
-                  final query = autocompleteQuery.query;
-                  return StreamMentionAutocompleteOptions(
-                    query: query,
-                    channel: StreamChannel.of(context).channel,
-                    mentionAllAppUsers: widget.mentionAllAppUsers,
-                    mentionsTileBuilder: widget.userMentionsTileBuilder,
-                    onMentionUserTap: (user) {
-                      // adding the mentioned user to the controller.
-                      _effectiveController.addMentionedUser(user);
+    return null;
+  }
 
-                      // accepting the autocomplete option.
-                      StreamAutocomplete.of(context)
-                          .acceptAutocompleteOption(user.name);
-                    },
-                  );
-                },
-              ),
-          ],
-        );
-      },
+  Widget? _buildDmCheckbox(BuildContext context) {
+    if (widget.hideSendAsDm) return null;
+
+    final insideThread = _effectiveController.message.parentId != null;
+    if (!insideThread) return null;
+
+    return DmCheckboxListTile(
+      value: _effectiveController.showInChannel,
+      contentPadding: const EdgeInsets.symmetric(horizontal: 8),
+      onChanged: (value) => _effectiveController.showInChannel = value,
+    );
+  }
+
+  Widget _buildNoPermissionMessage(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 24, vertical: 15),
+      child: Text(
+        context.translations.sendMessagePermissionError,
+        style: _messageInputTheme.inputTextStyle,
+      ),
     );
   }
 
@@ -1163,7 +1190,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
       hintType = HintType.searchGif;
     } else if (_effectiveController.attachments.isNotEmpty) {
       hintType = HintType.addACommentOrSend;
-    } else if (_effectiveController.cooldownTimeOut > 0) {
+    } else if (_effectiveController.isSlowModeActive) {
       hintType = HintType.slowModeOn;
     } else {
       hintType = HintType.writeAMessage;
@@ -1349,11 +1376,8 @@ class StreamMessageInputState extends State<StreamMessageInput>
 
   /// Sends the current message
   Future<void> sendMessage() async {
-    if (_effectiveController.cooldownTimeOut > 0 ||
-        (_effectiveController.text.trim().isEmpty &&
-            _effectiveController.attachments.isEmpty)) {
-      return;
-    }
+    if (_effectiveController.isSlowModeActive) return;
+    if (!widget.validator(_effectiveController.message)) return;
 
     final streamChannel = StreamChannel.of(context);
     final channel = streamChannel.channel;
@@ -1369,29 +1393,19 @@ class StreamMessageInputState extends State<StreamMessageInput>
           color: StreamChatTheme.of(context).colorTheme.accentError,
           size: 24,
         ),
-        title: 'Links are disabled',
-        details: 'Sending links is not allowed in this conversation.',
+        title: context.translations.linkDisabledError,
+        details: context.translations.linkDisabledDetails,
         okText: context.translations.okLabel,
       );
       return;
     }
 
-    final containsCommand = message.command != null;
-    // If the message contains command we should append it to the text
-    // before sending it.
-    if (containsCommand) {
-      message = message.copyWith(text: '/${message.command} ${message.text}');
-    }
-
-    var shouldKeepFocus = widget.shouldKeepFocusAfterMessage;
-    shouldKeepFocus ??= !_commandEnabled;
-
+    _maybeDeleteDraftMessage(message);
     widget.onQuotedMessageCleared?.call();
-
     _effectiveController.reset();
 
-    if (widget.preMessageSending != null) {
-      message = await widget.preMessageSending!(message);
+    if (widget.preMessageSending case final onPreMessageSending?) {
+      message = await onPreMessageSending.call(message);
     }
 
     // If the channel is not up to date, we should reload it before sending
@@ -1407,7 +1421,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
     await _sendOrUpdateMessage(message: message);
 
     if (mounted) {
-      if (shouldKeepFocus) {
+      if (widget.shouldKeepFocusAfterMessage ?? !_commandEnabled) {
         FocusScope.of(context).requestFocus(_effectiveFocusNode);
       } else {
         FocusScope.of(context).unfocus();
@@ -1460,6 +1474,56 @@ class StreamMessageInputState extends State<StreamMessageInput>
     );
   }
 
+  void _maybeUpdateOrDeleteDraftMessage() {
+    final message = _effectiveController.message;
+    final isMessageValid = widget.validator.call(message);
+
+    // If the message is valid, we need to create or update it as a draft
+    // message for the channel or thread.
+    if (isMessageValid) return _maybeUpdateDraftMessage(message);
+
+    // Otherwise, we need to delete the draft message.
+    return _maybeDeleteDraftMessage(message);
+  }
+
+  void _maybeUpdateDraftMessage(Message message) {
+    final channel = StreamChannel.of(context).channel;
+    final draft = switch (message.parentId) {
+      final parentId? => channel.state?.threadDraft(parentId),
+      null => channel.state?.draft,
+    };
+
+    final draftMessage = message.toDraftMessage();
+
+    // If the draft message didn't change, we don't need to update it.
+    if (draft?.message == draftMessage) return;
+
+    return channel.createDraft(draftMessage).ignore();
+  }
+
+  void _maybeDeleteDraftMessage(Message message) {
+    final channel = StreamChannel.of(context).channel;
+    final draft = switch (message.parentId) {
+      final parentId? => channel.state?.threadDraft(parentId),
+      null => channel.state?.draft,
+    };
+
+    // If there is no draft message, we don't need to delete it.
+    if (draft == null) return;
+
+    return channel.deleteDraft(parentId: message.parentId).ignore();
+  }
+
+  @override
+  void deactivate() {
+    final config = StreamChatConfiguration.of(context);
+    if (!_isEditing && config.draftMessagesEnabled) {
+      _maybeUpdateOrDeleteDraftMessage();
+    }
+
+    super.deactivate();
+  }
+
   @override
   void dispose() {
     _effectiveController.removeListener(_onChangedDebounced);
@@ -1468,6 +1532,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
     _focusNode?.dispose();
     _onChangedDebounced.cancel();
     _audioRecorderController.dispose();
+    _draftStreamSubscription?.cancel();
     super.dispose();
   }
 }