+++ packages/stream_chat_flutter_core/lib/src/stream_message_input_controller.dart
@@ -2,7 +2,7 @@ import 'dart:async' show Timer;
 import 'dart:convert';
 
 import 'package:collection/collection.dart';
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:stream_chat/stream_chat.dart';
 
 import 'package:stream_chat_flutter_core/src/message_text_field_controller.dart';
@@ -111,6 +111,9 @@ class StreamMessageInputController extends ValueNotifier<Message> {
     _textFieldController.text = text;
   }
 
+  /// Returns true if the slow mode is currently active.
+  bool get isSlowModeActive => _cooldownTimeOut > 0;
+
   /// The current [cooldownTimeOut] of the slow mode.
   ///
   /// Defaults to 0, which means slow mode is not active.
@@ -156,7 +159,7 @@ class StreamMessageInputController extends ValueNotifier<Message> {
   TextSelection get selection => _textFieldController.selection;
 
   set selection(TextSelection newSelection) {
-    _textFieldController.selection = selection;
+    _textFieldController.selection = newSelection;
   }
 
   /// Returns the textEditingValue associated with this controller.