+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_es.dart
@@ -560,10 +560,17 @@ No es posible añadir más de $limit archivos adjuntos
   String get updateYourCommentLabel => 'Actualizar tu comentario';
 
   @override
-  String get enterYourCommentLabel => 'Ingresar tu comentario';
+  String get enterYourCommentLabel => 'Ingresa tu comentario';
+
+  @override
+  String get endVoteConfirmationText =>
+      '¿Estás seguro de que quieres finalizar la votación?';
 
   @override
   String get createLabel => 'Crear';
+
+  @override
+  String get endLabel => 'Finalizar';
 
   @override
   String pollVotingModeLabel(PollVotingMode votingMode) {
@@ -669,4 +676,7 @@ No es posible añadir más de $limit archivos adjuntos
 
   @override
   String pollSomeoneCreatedText(String username) => '$username ha creado';
+
+  @override
+  String get draftLabel => 'Borrador';
 }