+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -562,10 +562,17 @@ Limite de pièces jointes dépassée : il n'est pas possible d'ajouter plus de $
   String get updateYourCommentLabel => 'Mettre à jour votre commentaire';
 
   @override
-  String get enterYourCommentLabel => 'Saisir votre commentaire';
+  String get enterYourCommentLabel => 'Entrez votre commentaire';
+
+  @override
+  String get endVoteConfirmationText =>
+      'Êtes-vous sûr de vouloir terminer le vote?';
 
   @override
   String get createLabel => 'Créer';
+
+  @override
+  String get endLabel => 'Terminer';
 
   @override
   String pollVotingModeLabel(PollVotingMode votingMode) {
@@ -672,4 +679,7 @@ Limite de pièces jointes dépassée : il n'est pas possible d'ajouter plus de $
 
   @override
   String pollSomeoneCreatedText(String username) => '$username a créé';
+
+  @override
+  String get draftLabel => 'Brouillon';
 }