+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_ko.dart
@@ -540,7 +540,13 @@ class StreamChatLocalizationsKo extends GlobalStreamChatLocalizations {
   String get enterYourCommentLabel => '댓글 입력';
 
   @override
+  String get endVoteConfirmationText => '투표를 종료하시겠습니까?';
+
+  @override
   String get createLabel => '생성';
+
+  @override
+  String get endLabel => '종료';
 
   @override
   String pollVotingModeLabel(PollVotingMode votingMode) {
@@ -642,4 +648,7 @@ class StreamChatLocalizationsKo extends GlobalStreamChatLocalizations {
 
   @override
   String pollSomeoneCreatedText(String username) => '$username님이 생성했습니다';
+
+  @override
+  String get draftLabel => '임시글';
 }