+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_pt.dart
@@ -561,7 +561,14 @@ Não é possível adicionar mais de $limit arquivos de uma vez
   String get enterYourCommentLabel => 'Inserir seu comentário';
 
   @override
+  String get endVoteConfirmationText =>
+      'Tem certeza de que deseja encerrar a votação?';
+
+  @override
   String get createLabel => 'Criar';
+
+  @override
+  String get endLabel => 'Encerrar';
 
   @override
   String pollVotingModeLabel(PollVotingMode votingMode) {
@@ -666,4 +673,7 @@ Não é possível adicionar mais de $limit arquivos de uma vez
 
   @override
   String pollSomeoneCreatedText(String username) => '$username criou';
+
+  @override
+  String get draftLabel => 'Rascunho';
 }