+++ packages/stream_chat/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 9.9.0
+
+✅ Added
+
+- Added teams role support for users.
+- Added support for Filtering and Sorting in the `client.queryThreads` method.
+
 ## 9.8.0
 
 ✅ Added